feat: resuelto el problema de la API en el frontend

+++ src/main/java/com/MyInventory/dto/DashboardResponse.java
@@ -19,5 +19,5 @@ public class DashboardResponse {
 
     private Long totalExits;
 
-    private Integer totalStock;
+    private Long totalStock;
 }
